pyproject.toml
[project]
authors = [
  { name = "Björn F. Johansson", email = "bjornjobb@gmail.com" },
  { name = "Manuel Lera-Ramirez", email = "manulera14@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Education",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Education",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]
# description = "Representing double stranded DNA and functions for simulating cloning and homologous recombination between DNA molecules."
dynamic = ["version"]
name = "pydna"
readme = "README.md"
[tool.poetry.urls]
Changelog = "https://github.com/pydna-group/pydna/blob/master/docs/CHANGELOG.md#changelog"
[tool.poetry]
description = "Representing double stranded DNA and functions for simulating cloning and homologous recombination between DNA molecules."
authors = [
  "Björn F. Johansson <bjornjobb@gmail.com>",
  "Manuel Lera-Ramirez <manulera14@gmail.com>",
]
documentation = "https://pydna-group.github.io/pydna"
homepage = "https://github.com/pydna-group/pydna#-pydna"
license = "BSD"
name = "pydna"
readme = "README.md"
repository = "https://github.com/pydna-group/pydna/tree/master"
version = "6.0.0-a.24.post.17+b7b559bd66"
[tool.poetry.dependencies]
appdirs = ">=1.4.4"
biopython = "1.85"
cai2 = { version = ">=1.0.5", optional = true }
matplotlib = { version = ">=3.4.3", optional = true }
networkx = ">=2.8.8"
numpy = [
  { version = ">1.26", python = "<3.12" },
  { version = ">=2.3.0", python = ">=3.12" },
]
pillow = { version = ">=8.4.0", optional = true }
prettytable = ">=3.5.0"
pydivsufsort = ">=0.0.14"
pyfiglet = "0.8.post1"
pyparsing = { version = ">=2.4.7", optional = true }
pyperclip = { version = ">=1.8.2", optional = true }
python = ">=3.10,<4.0"
requests = { version = ">=2.26.0", optional = true }
pyahocorasick = { version = ">=2.2.0", optional = true }
scipy = [
  { version = ">=1.11.3", python = ">=3.12", optional = true },
  { version = ">=1.9.3", python = "<3.12", optional = true },
]
seguid = ">=0.0.5"
regex = "^2024.11.6"
opencloning-linkml = "^0.4.9"
[tool.poetry.extras]
clipboard = ["pyperclip"]
download = ["pyparsing", "requests"]
express = ["cai2"]
gel = ["scipy", "matplotlib", "pillow"]
primer_screen = ["pyahocorasick"]
[build-system]
build-backend = "poetry_dynamic_versioning.backend"
requires = ["poetry-core", "poetry-dynamic-versioning"]
[tool.poetry-dynamic-versioning]
enable = true
style = "semver"
vcs = "git"
[tool.poetry-dynamic-versioning.substitution]
files = ["src/pydna/__init__.py"]
[tool.poetry.group.test] # This part can be left out
[tool.poetry.group.test.dependencies]
coverage = ">=7.1.0"
nbval = ">=0.9.6"
pytest = ">=7.2.0"
pytest-cov = ">=4.0.0"
pytest-doctestplus = ">=0.12.1"
pytest-profiling = ">=1.7.0"
requests-mock = ">=1.10.0"

[tool.poetry.group.docs.dependencies]
numpydoc = "^1.6.0"
sphinx = "*"
# sphinx-rtd-theme = ">=1.3,<4.0"
# nbconvert = "^7.16.4"
# myst-parser = [
#   { version = ">3.0.0", python = "<3.10" },
#   { version = ">=4.0.0", python = ">=3.10" },
# ]
myst-nb = "*"
sphinx-book-theme = "*"
pandas = "*"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.1.0"
black = "^25.1.0"
pre-commit = ">3.6"
nbstripout = "^0.7.1"
flake8-bugbear = "^24.12.12"

[tool.pytest.ini_options]
minversion = "6.0.2"
python_files = "test_*.py"
testpaths = ["tests", "src/pydna"]
norecursedirs = ["*"]

[tool.black]
include = '\.pyi?$'
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.9"
